+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.60.1] - 2024-12-11
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): API error handling for time preparation.
+- [Marek Kurańda](https://github.com/mjkuranda): Sorting ingredient labels which caused improper ID generation.
+- [Marek Kurańda](https://github.com/mjkuranda): Searching dishes providing types.
+
 ## [1.60.0] - 2024-11-27
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): New subpage `/about` where put information about the author.